requirements.txt
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
cryptography==44.0.2
distro==1.9.0
et_xmlfile==2.0.0
numpy==2.2.4
opencv-python-headless==4.11.0.86
openpyxl==3.1.5
pandas==2.2.3
pdfminer.six==20250324
pycparser==2.22
pypdf==5.4.0
pypdfium2==4.30.1
python-dateutil==2.9.0.post0
pytz==2025.2
six==1.17.0
tabula-py==2.10.0
tabulate==0.9.0
tzdata==2025.2
jpype1
